requirements.txt
asgiref==3.8.1
Django==5.1.4
django-extensions==3.2.3
djangorestframework==3.15.2
gunicorn==23.0.0
Markdown==3.7
markdown2==2.5.2
packaging==24.2
pillow==11.1.0
sqlparse==0.5.3
tzdata==2024.2
whitenoise==6.9.0
